Create workers.yml
The first play gets the join command that needs to be run on the worker nodes. This command will be in the following format:kubeadm join --token <token> <master-ip>:<master-port> --discovery-token-ca-cert-hash sha256:<hash>. Once it gets the actual command with the proper token and hash values, the task sets it as a fact so that the next play will be able to access that info.
The second play has a single task that runs the join command on all worker nodes. On completion of this task, the two worker nodes will be part of the cluster.